backend/pyproject.toml
[project]
name = "pathology-voice-backend"
version = "0.1.0"
description = "FastAPI backend for Pathology Lab Voice Agent (Bolna + Supabase)"
requires-python = ">=3.10,<3.11"

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "httpx>=0.27.0",
  "asyncpg>=0.29.0",
  "supabase-py>=2.5.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

